# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Testing Swagger APIs.",
        "title": "Matchmaking Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "0.2"
    },
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Authenticate a user with their username and password and return a signed JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/inputs.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request or credentials",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Refresh the current authenticated user's session using the refresh token cookie or header.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh token",
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized or invalid refresh token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "Register a new user and create their default profile record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "User signup payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/inputs.AuthInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation or duplicate user error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/users/avatar/fetch": {
            "get": {
                "description": "Given an existing S3 object key, generates and returns a 7-day presigned\nGET URL for the corresponding object in the \"matchmaking\" bucket.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Get a presigned image URL",
                "parameters": [
                    {
                        "type": "string",
                        "example": "images/1234567890_photo.jpg",
                        "description": "S3 object key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Presigned URL generated successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.ImageUploadResponse"
                        }
                    },
                    "400": {
                        "description": "Missing key query parameter",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to generate presigned URL",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/avatar/upload": {
            "post": {
                "description": "Accepts a multipart/form-data request containing a single image file,\nvalidates its MIME type, stores it in the \"matchmaking\" S3 bucket under\nthe images/ prefix, and returns a 7-day presigned GET URL.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Upload an image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file to upload (JPEG, PNG, WebP, or GIF; max 10 MB)",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Upload successful — presigned URL valid for 7 days",
                        "schema": {
                            "$ref": "#/definitions/handlers.ImageUploadResponse"
                        }
                    },
                    "400": {
                        "description": "Request too large, not multipart, or missing the 'image' field",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "415": {
                        "description": "MIME type not allowed",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnsupportedMediaTypeResponse"
                        }
                    },
                    "500": {
                        "description": "S3 upload failed or presign failed",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "missing 'image' field"
                }
            }
        },
        "handlers.ImageUploadResponse": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string",
                    "example": "images/1234567890_photo.jpg"
                },
                "url": {
                    "type": "string",
                    "example": "https://s3.example.com/matchmaking/images/1234567890_photo.jpg?X-Amz-Signature=..."
                }
            }
        },
        "handlers.UnsupportedMediaTypeResponse": {
            "type": "object",
            "properties": {
                "allowed": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "image/jpeg",
                        "image/png",
                        "image/webp",
                        "image/gif"
                    ]
                },
                "error": {
                    "type": "string",
                    "example": "unsupported image type"
                }
            }
        },
        "inputs.AuthInput": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "inputs.LoginInput": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}